animal/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
authlib>=1.2.1
requests>=2.31.0
jwt>=1.3.1
python-jose
six
protobuf==3.20.*